[cheri] Adapt external entity table protections.
Instances of the external entity table are used to store capability
values, such as the external pointer table. To write capabilitiy values
to the table the mmap protections of the allocated pages must ensure the
correct access. Ideally this would be performed by setting the PROT_MAX
value to allow setting the permissions after the initial mapping.
However, as this is not current possible the initial mapping is created
with kReadWrite permissions.

Support for passing the maximum permissions has been added to allow
creating the mapping as kNoAccess and changing the permissions when this
is supported in CheriBSD.