Simple example of restricting DDC for isolates
This doesn't work _but_ it's useful: demonstrates that we immediately
hit similar problems when restricting isolates that we do when
restricting function calls (it's really hard to find a reasonable set of
bounds!)

Laurie suggested moving the call stack inside the isolate to make this
easier. It'd require a bit of a change to where our restriction is
inserted, but I'm certain we could make this work... I wonder whether
it'd be more pragmatic to restrict bounds that way. And would it help
get security benefits? What about e.g. the heap? Can we avoid corruption
there or do we need to do extra work to protect it?